Add patch from upstream to prevent contained apps from using TIOCSTI ioctl
This would let the app inject commands into the terminal from which it
was invoked (CVE-2017-5226). This was initially fixed in bubblewrap by
calling setsid(), but that breaks the ability to use Ctrl+Z or Ctrl+C
on a flatpak-confined process, so it is being made optional; prevent
the attack here instead, in a way that doesn't break shells.