Skip parental controls if accountsservice is not installed
The malcontent package (which activates parental controls support)
depends on accountsservice, but the libmalcontent-0-0 client library
does not, so we need to cope gracefully with the case where
neither malcontent nor accountsservice is installed. Presumably, in such
installations the sysadmin did not want the parental controls feature.

Ideally libmalcontent would do this itself (#972145).

Closes: #972138