d/p/system-helper-Fix-deploys-of-local-remotes.patch: Add
Fix some failures to update in GNOME Software and the unit tests.

This change was previously applied in Ubuntu's flatpak_1.10.2-1ubuntu1
to fix a unit test failure, possibly triggered by a newer version of
GLib. It has also been reported to fix a failure to upgrade Flatpak
apps using GNOME Software, this time in Fedora.

`git show --ignore-space-change c8b9069a` shows a version of this patch
that will be easier to review.