d/p/create-usb-Skip-copying-extra-data-flatpaks.patch: Add
Skip flatpaks with "extra-data" when using `flatpak create-usb`.

This command is intended to create USB drives that can be used to install
Flatpak apps and/or runtimes while offline, but the "extra-data" feature
downloads extra content for an app or runtime at install time, as a way
to automate installation of data that can be re-downloaded by end users
but is not licensed for redistribution by Flatpak repositories. Such
apps and runtimes would fail to install while offline.

This is not a critical change for Debian in itself, but it's low-risk,
and its inclusion makes this package match what would be in Flatpak
stable release 1.10.3 if it was released today.