d/rules: Tighten up handling of nocheck and noinsttest
The upstream build system checks for some programs that are required
during testing whenever the tests are compiled. If we are under both
the nocheck and noinsttest build profiles, don't compile the tests,
so that pkcheck and socat won't be needed in that configuration;
and otherwise, we need them in Build-Depends.

Closes: #1116737